+++ openff/interchange/interop/parmed.py
@@ -18,16 +18,16 @@ from openff.interchange.models import (
     ProperTorsionKey,
     TopologyKey,
 )
+from openff.interchange.smirnoff._valence import (
+    SMIRNOFFImproperTorsionCollection,
+    SMIRNOFFProperTorsionCollection,
+)
 
 if TYPE_CHECKING:
 
     import parmed as pmd
 
     from openff.interchange import Interchange
-    from openff.interchange.components.smirnoff import (
-        SMIRNOFFImproperTorsionHandler,
-        SMIRNOFFProperTorsionHandler,
-    )
 
 kcal_mol = unit.Unit("kilocalories / mol")
 kcal_mol_a2 = unit.Unit("kilocalories / mol / angstrom ** 2")
@@ -288,13 +288,15 @@ def _from_parmed(cls, structure) -> "Interchange":
     else:
         raise ConversionError("ParmEd Structure missing an topology attribute")
 
-    from openff.interchange.components.smirnoff import (
-        SMIRNOFFAngleHandler,
-        SMIRNOFFBondHandler,
-        SMIRNOFFElectrostaticsHandler,
-        SMIRNOFFImproperTorsionHandler,
-        SMIRNOFFProperTorsionHandler,
-        SMIRNOFFvdWHandler,
+    from openff.interchange.smirnoff._nonbonded import (
+        SMIRNOFFElectrostaticsCollection,
+        SMIRNOFFvdWCollection,
+    )
+    from openff.interchange.smirnoff._valence import (
+        SMIRNOFFAngleCollection,
+        SMIRNOFFBondCollection,
+        SMIRNOFFImproperTorsionCollection,
+        SMIRNOFFProperTorsionCollection,
     )
 
     _scale_14_coul = {1 / d.scee for d in structure.dihedral_types if d.scee != 0}
@@ -316,8 +318,11 @@ def _from_parmed(cls, structure) -> "Interchange":
         scale_14_vdw = [*_scale_14_vdw][0]
 
     # TODO: Infer 1-4 scaling factors from exceptions/adjusts/dihedrals/something
-    vdw_handler = SMIRNOFFvdWHandler(scale_14=scale_14_vdw)
-    coul_handler = SMIRNOFFElectrostaticsHandler(scale_14=scale_14_coul, method="pme")
+    vdw_handler = SMIRNOFFvdWCollection(scale_14=scale_14_vdw)
+    coul_handler = SMIRNOFFElectrostaticsCollection(
+        scale_14=scale_14_coul,
+        method="pme",
+    )
 
     for atom in structure.atoms:
         atom_idx = atom.idx
@@ -336,7 +341,7 @@ def _from_parmed(cls, structure) -> "Interchange":
             {pot_key: Potential(parameters={"charge": charge})},
         )
 
-    bond_handler = SMIRNOFFBondHandler()
+    bond_handler = SMIRNOFFBondCollection()
 
     for bond in structure.bonds:
         atom1 = bond.atom1
@@ -354,7 +359,7 @@ def _from_parmed(cls, structure) -> "Interchange":
     out.collections.update({"Electrostatics": coul_handler})
     out.collections.update({"Bonds": bond_handler})
 
-    angle_handler = SMIRNOFFAngleHandler()
+    angle_handler = SMIRNOFFAngleCollection()
 
     for angle in structure.angles:
         atom1 = angle.atom1
@@ -369,8 +374,8 @@ def _from_parmed(cls, structure) -> "Interchange":
         angle_handler.slot_map.update({top_key: pot_key})
         angle_handler.potentials.update({pot_key: pot})
 
-    proper_torsion_handler = SMIRNOFFProperTorsionHandler()
-    improper_torsion_handler = SMIRNOFFImproperTorsionHandler()
+    proper_torsion_handler = SMIRNOFFProperTorsionCollection()
+    improper_torsion_handler = SMIRNOFFImproperTorsionCollection()
 
     for dihedral in structure.dihedrals:
         if isinstance(dihedral.type, pmd.DihedralType):
@@ -435,7 +440,10 @@ def _lj_params_from_potential(potential):
 def _process_single_dihedral(
     dihedral: "pmd.Dihedral",
     dihedral_type: "pmd.DihedralType",
-    handler: Union["SMIRNOFFImproperTorsionHandler", "SMIRNOFFProperTorsionHandler"],
+    collection: Union[
+        SMIRNOFFImproperTorsionCollection,
+        SMIRNOFFProperTorsionCollection,
+    ],
     mult: Optional[int] = None,
 ):
     atom1 = dihedral.atom1
@@ -461,13 +469,13 @@ def _process_single_dihedral(
         )
         pot = Potential(parameters={"k": k, "periodicity": periodicity, "phase": phase})
 
-        if pot_key in handler.potentials:
+        if pot_key in collection.potentials:
             raise Exception(
                 "This dihedral already exists, indices are probably messed up.",
             )
 
-        handler.slot_map.update({improper_key: pot_key})
-        handler.potentials.update({pot_key: pot})
+        collection.slot_map.update({improper_key: pot_key})
+        collection.potentials.update({pot_key: pot})
     else:
         proper_key = ProperTorsionKey(
             atom_indices=(atom1.idx, atom2.idx, atom3.idx, atom4.idx),
@@ -479,9 +487,9 @@ def _process_single_dihedral(
         )
         pot = Potential(parameters={"k": k, "periodicity": periodicity, "phase": phase})
 
-        while pot_key in handler.potentials:
+        while pot_key in collection.potentials:
             pot_key.mult += 1  # type: ignore
             proper_key.mult += 1  # type: ignore
 
-        handler.slot_map.update({proper_key: pot_key})
-        handler.potentials.update({pot_key: pot})
+        collection.slot_map.update({proper_key: pot_key})
+        collection.potentials.update({pot_key: pot})
